Bloogy  V.0.4.2.2
Final report functionalities integration

New views for comment admin index with the posibility to chose beetween All the Comment or just the Reported Comments

New Function -> getReportedComments in CommentTable  -> Get the comments and the count of report when it'is reported.

In admin.comment.index.php

A condition is using to show or not the comment in Reported Comments tab -> if the count of report for a comment is more than zero, it is showed, if not, not showed.

$reportedComments variable introduce in CommentsController